update for pr feedback
