+++ components/nimbus/src/tests/stateful/test_evaluator.rs
@@ -9,9 +9,10 @@ use crate::{
     },
     enrollment::NotEnrolledReason,
     evaluator::targeting,
-    EnrollmentStatus,
+    EnrollmentStatus, TargetingAttributes, AppContext,
 };
 use chrono::Utc;
+use std::collections::HashSet;
 
 #[test]
 fn test_event_sum_transform() {
@@ -259,6 +260,105 @@ fn test_event_transform_last_seen_parameters() {
         Some(EnrollmentStatus::Error {
             reason: "EvaluationError: Custom error: Transform parameter error: event transform LastSeen requires a positive number as the second parameter"
                 .to_string()
+        })
+    );
+}
+
+#[test]
+fn test_targeting_active_experiments_equivalency() {
+    // Here's our valid jexl statement
+    let expression_statement = "'test' in active_experiments";
+    // A matching context that includes the appropriate specific context
+    let mut targeting_attributes: TargetingAttributes = AppContext {
+        app_name: "nimbus_test".to_string(),
+        app_id: "1010".to_string(),
+        channel: "test".to_string(),
+        app_version: Some("4.4".to_string()),
+        app_build: Some("1234".to_string()),
+        architecture: Some("x86_64".to_string()),
+        device_manufacturer: Some("Samsung".to_string()),
+        device_model: Some("Galaxy S10".to_string()),
+        locale: Some("en-US".to_string()),
+        os: Some("Android".to_string()),
+        os_version: Some("10".to_string()),
+        android_sdk_version: Some("29".to_string()),
+        debug_tag: None,
+        custom_targeting_attributes: None,
+        ..Default::default()
+    }
+        .into();
+    let mut set = HashSet::<String>::new();
+    set.insert("test".into());
+    targeting_attributes.active_experiments = set;
+
+    // The targeting should pass!
+    assert_eq!(
+        targeting(expression_statement, &targeting_attributes.clone().into()),
+        None
+    );
+
+    // We set active_experiment treatment to something not expected and try again
+    let mut set = HashSet::<String>::new();
+    set.insert("test1".into());
+    targeting_attributes.active_experiments = set;
+    assert_eq!(
+        targeting(expression_statement, &targeting_attributes.clone().into()),
+        Some(EnrollmentStatus::NotEnrolled {
+            reason: NotEnrolledReason::NotTargeted
+        })
+    );
+
+    // We set active_experiments to None and try again
+    let set = HashSet::<String>::new();
+    targeting_attributes.active_experiments = set;
+    assert_eq!(
+        targeting(expression_statement, &targeting_attributes.into()),
+        Some(EnrollmentStatus::NotEnrolled {
+            reason: NotEnrolledReason::NotTargeted
+        })
+    );
+}
+
+#[test]
+fn test_targeting_active_experiments_exists() {
+    // Here's our valid jexl statement
+    let expression_statement = "'test' in active_experiments";
+    // A matching context that includes the appropriate specific context
+    let mut targeting_attributes: TargetingAttributes = AppContext {
+        app_name: "nimbus_test".to_string(),
+        app_id: "1010".to_string(),
+        channel: "test".to_string(),
+        app_version: Some("4.4".to_string()),
+        app_build: Some("1234".to_string()),
+        architecture: Some("x86_64".to_string()),
+        device_manufacturer: Some("Samsung".to_string()),
+        device_model: Some("Galaxy S10".to_string()),
+        locale: Some("en-US".to_string()),
+        os: Some("Android".to_string()),
+        os_version: Some("10".to_string()),
+        android_sdk_version: Some("29".to_string()),
+        debug_tag: None,
+        custom_targeting_attributes: None,
+        ..Default::default()
+    }
+        .into();
+    let mut set = HashSet::<String>::new();
+    set.insert("test".into());
+    targeting_attributes.active_experiments = set;
+
+    // The targeting should pass!
+    assert_eq!(
+        targeting(expression_statement, &targeting_attributes.clone().into()),
+        None
+    );
+
+    // We set active_experiment treatment to something not expected and try again
+    let set = HashSet::<String>::new();
+    targeting_attributes.active_experiments = set;
+    assert_eq!(
+        targeting(expression_statement, &targeting_attributes.into()),
+        Some(EnrollmentStatus::NotEnrolled {
+            reason: NotEnrolledReason::NotTargeted
         })
     );
 }
